@@ -1,35 +1,61 @@
 {
-    "name": "topthink/think",
-    "description": "the new thinkphp framework",
+    "name": "coolxitech/dnsmgr",
+    "description": "聚合DNS管理系统",
+    "version": "1.7.2",
     "type": "project",
     "keywords": [
-        "framework",
         "thinkphp",
-        "ORM"
+        "dns",
+        "dnsmanager",
+        "cccyun"
     ],
-    "homepage": "https://www.thinkphp.cn/",
-    "license": "Apache-2.0",
+    "homepage": "https://blog.cccyun.cn/post-526.html",
+    "license": "MIT",
     "authors": [
         {
             "name": "liu21st",
-            "email": "liu21st@gmail.com"
+            "email": "liu21st@gmail.com",
+            "role": "Framework Developer"
         },
         {
             "name": "yunwuxin",
-            "email": "448901948@qq.com"
-        }        
+            "email": "448901948@qq.com",
+            "role": "Framework Developer"
+        },
+        {
+            "name": "netcccyun",
+            "homepage": "https://blog.cccyun.cn",
+            "role": "Project Owner"
+        },
+        {
+            "name": "coolxitech",
+            "email": "admin@kuxi.tech",
+            "homepage": "https://www.kuxi.tech",
+            "role": "Project Developer"
+        }
     ],
+    "support": {
+        "issues": "https://github.com/coolxitech/dnsmgr/issues",
+        "source": "https://github.com/coolxitech/dnsmgr"
+    },
     "require": {
-        "php": ">=7.2.5",
-        "topthink/framework": "^6.0.0",
-        "topthink/think-orm": "^2.0",
-        "topthink/think-view": "^1.0",
-        "cccyun/think-captcha": "^3.0"
+        "php": ">=8.0.0",
+        "ext-pdo": "*",
+        "ext-gd": "*",
+        "ext-curl": "*",
+        "topthink/framework": "^8.0",
+        "topthink/think-orm": "^3.0",
+        "topthink/think-filesystem": "^2.0",
+        "topthink/think-view": "^2.0",
+        "topthink/think-captcha": "^3.0",
+        "guzzlehttp/guzzle": "^7.9",
+        "coolxitech/cloudflare-sdk": "dev-master"
     },
     "require-dev": {
         "symfony/var-dumper": "^4.2",
         "topthink/think-trace":"^1.0",
-        "swoole/ide-helper": "^5.1"
+        "swoole/ide-helper": "^5.1",
+        "friendsofphp/php-cs-fixer": "^3.64.0"
     },
     "autoload": {
         "psr-4": {
@@ -47,5 +73,13 @@
             "@php think service:discover",
             "@php think vendor:publish"
         ]
-    }
+    },
+    "repositories": [
+        {
+            "description": "阿里云镜像",
+            "type": "composer",
+            "url": "https://mirrors.aliyun.com/composer/",
+            "canonical": false
+        }
+    ]
 }
