@@ -1,8 +1,8 @@
 <?php
+
 namespace app;
 
 // 应用请求对象类
 class Request extends \think\Request
 {
-
 }
