@@ -1,4 +1,5 @@
 <?php
+
 declare (strict_types=1);
 
 namespace app\middleware;
@@ -12,28 +13,28 @@ class AuthUser
         $islogin = false;
         $cookie = cookie('user_token');
         $user = null;
-        if($cookie && config_get('sys_key')){
-            $token=authcode($cookie, 'DECODE', config_get('sys_key'));
-            if($token){
+        if ($cookie && config_get('sys_key')) {
+            $token = authcode($cookie, 'DECODE', config_get('sys_key'));
+            if ($token) {
                 list($type, $uid, $sid, $expiretime) = explode("\t", $token);
-                if($type == 'user'){
+                if ($type == 'user') {
                     $user = Db::name('user')->where('id', $uid)->find();
-                    if($user && $user['status']==1){
-                        $session=md5($user['id'].$user['password']);
-                        if($session==$sid && $expiretime>time()) {
+                    if ($user && $user['status'] == 1) {
+                        $session = md5($user['id'].$user['password']);
+                        if ($session == $sid && $expiretime > time()) {
                             $islogin = true;
                         }
                         $user['type'] = 'user';
                         $user['permission'] = [];
-                        if($user['level'] == 1){
+                        if ($user['level'] == 1) {
                             $user['permission'] = Db::name('permission')->where('uid', $uid)->column('domain');
                         }
                     }
-                }elseif($type == 'domain'){
+                } elseif ($type == 'domain') {
                     $user = Db::name('domain')->where('id', $uid)->find();
-                    if($user && $user['is_sso']==1){
-                        $session=md5($user['id'].$user['name']);
-                        if($session==$sid && $expiretime>time()) {
+                    if ($user && $user['is_sso'] == 1) {
+                        $session = md5($user['id'].$user['name']);
+                        if ($session == $sid && $expiretime > time()) {
                             $islogin = true;
                         }
                         $user['username'] = $user['name'];
