@@ -1,16 +1,16 @@
 <?php
+
 declare (strict_types=1);
 
 namespace app\middleware;
-
 
 class CheckLogin
 {
     public function handle($request, \Closure $next)
     {
-        if (!request()->islogin) {
+        if (!$request->islogin) {
             if ($request->isAjax() || !$request->isGet()) {
-                return json(['code'=>-1, 'msg'=>'未登录'])->code(401);
+                return json(['code' => -1, 'msg' => '未登录'])->code(401);
             }
             return redirect((string)url('/login'));
         }
