@@ -7,7 +7,7 @@ return [
     //验证码位数
     'length'   => 4,
     // 验证码字符集合
-    'codeSet'  => '2345678abcdefhijkmnpqrstuvwxyzABCDEFGHJKLMNPQRTUVWXY',
+    'codeSet'  => '1234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ',
     // 验证码过期时间
     'expire'   => 1800,
     // 是否使用中文验证码
@@ -30,7 +30,8 @@ return [
     'imageH'   => 0,
     // 验证码图片宽度
     'imageW'   => 0,
-
+    // 验证成功后是否重置
+    'reset' => true,
     // 添加额外的验证码设置
     // verify => [
     //     'length'=>4,
